+++ Deno.novaextension/Scripts/Queue.js
@@ -1,3 +1,5 @@
+// @ts-check
+
 /**
  * A task queue runner that runs added tasks after all previously added tasks
  * have settled.
@@ -9,8 +11,9 @@ class Queue {
 
   /**
    * Adds a task to the queue.
-   * @param {function} task Task, that may return a promise.
-   * @returns {Promise} Resolves what the task returns.
+   * @template [T = void | Promise<void>]
+   * @param {() => T} task Task, that may return a promise.
+   * @returns {Promise<T>} Resolves what the task returns.
    */
   // deno-lint-ignore require-await
   async add(task) {
@@ -18,11 +21,13 @@ class Queue {
       throw new TypeError("Argument 1 `task` must be a function.");
     }
 
-    const promise = Promise.allSettled(this.queue).then(task).finally(() => {
-      // Remove the settled queue item to allow garbage collection and prevent
-      // the queue from growing infinitely long.
-      this.queue.delete(promise);
-    });
+    const promise = Promise
+      // @ts-ignore This seems to work in the Nova JavaScript environment.
+      .allSettled(this.queue).then(task).finally(() => {
+        // Remove the settled queue item to allow garbage collection and prevent
+        // the queue from growing infinitely long.
+        this.queue.delete(promise);
+      });
 
     this.queue.add(promise);