+++ Deno.novaextension/Scripts/restartLanguageClient.js
@@ -1,11 +1,17 @@
+// @ts-check
+/// <reference path="https://unpkg.com/@types/nova-editor-node@4.1.4/index.d.ts" />
+
+/** @type {typeof import("./Queue.js")} */
 const Queue = require("./Queue.js");
+
+/** @type {typeof import("./stopLanguageClient.js")} */
 const stopLanguageClient = require("./stopLanguageClient.js");
 
 const restartLanguageClientQueue = new Queue();
 
 /**
  * Restarts a Nova language client, making sure multiple restarts queue.
- * @param {LanguageClient} The Nova language client to restart.
+ * @param {LanguageClient} languageClient The Nova language client to restart.
  * @returns {Promise<void>} Resolves once the restart finishes.
  */
 async function restartLanguageClient(languageClient) {