+++ Deno.novaextension/Scripts/setNovaGlobalConfigMapEntry.js
@@ -1,10 +1,14 @@
+// @ts-check
+/// <reference path="https://unpkg.com/@types/nova-editor-node@4.1.4/index.d.ts" />
+
+/** @type {typeof import("./setNovaGlobalConfig.js")} */
 const setNovaGlobalConfig = require("./setNovaGlobalConfig.js");
 
 /**
  * Sets a Nova global config map entry, creating the map if it doesn’t exist.
  * @param {string} configKey Nova global config key.
  * @param {string} mapKey Map key.
- * @param {Any} mapValue Map value.
+ * @param {unknown} mapValue Map value.
  */
 async function setNovaGlobalConfigMapEntry(configKey, mapKey, mapValue) {
   if (typeof configKey !== "string") {